Implementing spring security on backend.

+++ backend/src/main/java/com/appslab/CloudService/Controllers/UploadFileController.java
@@ -2,6 +2,7 @@ package com.appslab.CloudService.Controllers;
 
 import com.appslab.CloudService.Services.UploadFileService;
 import com.appslab.CloudService.Models.UploadedFile;
+import com.appslab.CloudService.Services.UserService;
 import org.springframework.core.io.FileSystemResource;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.data.repository.query.Param;
@@ -16,9 +17,11 @@ import java.util.List;
 @RestController
 public class UploadFileController {
     UploadFileService uploadFileService;
+    UserService userService;
 
-    public UploadFileController(UploadFileService uploadFileService) {
+    public UploadFileController(UploadFileService uploadFileService, UserService userService) {
         this.uploadFileService = uploadFileService;
+        this.userService = userService;
     }
 
     @PostMapping
@@ -32,25 +35,34 @@ public class UploadFileController {
 
     @GetMapping
     public List<UploadedFile> listOfFiles(){
-        return uploadFileService.listOfFiles();
+        Long specifyUserId = userService.getSpecifyUserId();
+        return uploadFileService.listOfFiles(specifyUserId);
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<InputStreamResource> getFile(@PathVariable Long id) throws Exception{
         UploadedFile uploadedFile = uploadFileService.findFileById(id).get();
-        FileSystemResource file = new FileSystemResource(uploadFileService.pathToSpecificFile(uploadedFile));
-        return ResponseEntity.ok().contentType(MediaType.parseMediaType(uploadedFile.getMimeType())).body(new InputStreamResource(file.getInputStream()));
+        if(uploadedFile.getCustomUserId()==userService.getSpecifyUserId())
+        {
+            FileSystemResource file = new FileSystemResource(uploadFileService.pathToSpecificFile(uploadedFile));
+            return ResponseEntity.ok().contentType(MediaType.parseMediaType(uploadedFile.getMimeType())).body(new InputStreamResource(file.getInputStream()));
+        }
+        return null;
     }
 
     @DeleteMapping("/{id}")
     public void deleteFile(@PathVariable Long id) throws Exception{
         UploadedFile uploadedFile = uploadFileService.findFileById(id).get();
-        Files.delete(uploadFileService.pathToSpecificFile(uploadedFile));
-        uploadFileService.deleteFile(id);
+        if (uploadedFile.getCustomUserId()== userService.getSpecifyUserId())
+        {
+            Files.delete(uploadFileService.pathToSpecificFile(uploadedFile));
+            uploadFileService.deleteFile(id);
+        }
     }
 
     @GetMapping("/search/")
     public List<UploadedFile> uploadedFile(@Param("keyword") String keyword){
-        return uploadFileService.findSearchingFiles(keyword);
+        Long specifiUserId = userService.getSpecifyUserId();
+        return uploadFileService.findSearchingFiles(keyword,specifiUserId);
     }
 }
